database column fixes and form modifications

+++ db/migrate/20160318201830_create_ticket_details.rb
@@ -1,8 +1,8 @@
 class CreateTicketDetails < ActiveRecord::Migration
   def change
     create_table :ticket_details do |t|
-      
-      t.decimal :price
+
+      t.integer :price
 
       t.timestamps null: false
     end
